move validation- and preprocessing-code to a separate preprocessor module.
