+++ src/libs/ontouml2alloy/index.ts
@@ -3,4 +3,5 @@ export * from './class_functions';
 export * from './generalization_functions';
 export * from './generalization_set_functions';
 export * from './relation_functions';
+export * from './preprocessor';
 export * from './ontouml2alloy';
